Added support for custom Java-based editor.
A custom Java-based editor can now be added to a JVstHost object. The view can modify parameter and program settings of the vst, as well as send midi messages to it. JVstHost has been modified such that all public methods are synchronized. This helps coordinate automation and ui threads which may try to access the plugin simultaneously.